+++ test/bonny/naming_test.exs
@@ -1,0 +1,5 @@
+defmodule Bonny.NamingTest do
+  @moduledoc false
+  use ExUnit.Case, async: true
+  doctest Bonny.Naming
+end